lib: performance improvements
Add a 'profiling' profile and make some initial changes based on the
results.

The big win: parsing JSON directly out of a zstd::Decoder turns out to
be pretty slow, even if we add a BufReader, so decompress it in one go
first and use the serde_json slice API on the result.

A smaller win: instead of using an iterator-based JSON parser for the
tar-split, just split on lines and parse each line separately.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>