examples: add a pull example
🦎

Use the oci-client crate to pull a zstd:chunked image using something
approximating a performant and robust downloader.

For now this just creates a "tmp" directory in the current directory and
puts the chunks there.  It doesn't actually do anything with the data
once it's downloaded, but at that point it would be easy to reassemble
it into some form.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>