lib: drop the idea of multiple chunks per file
For Chunk::External we can theoretically return multiple chunks for
files that have been subdivided (by the rollsum, for example).  I think
it's time to lay this idea for rest, at least for the time being, and
improve the ergonomics and efficiency of the API by replacing the
Box<[ContentReference]> with a single ContentReference.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>